test: organize refactor coverage by domain
