+++ CaskHub/Services/CaskActionPresentation.swift
@@ -2,6 +2,8 @@
 //  CaskActionPresentation.swift
 //  CaskHub
 //
+//  Created by Ali Elsokary on 25/07/2026.
+//
 
 nonisolated enum CaskActionAlert: Equatable, Sendable {
     case packageAdoption
