+++ CaskHub/Services/CaskOperationStore.swift
@@ -2,6 +2,8 @@
 //  CaskOperationStore.swift
 //  CaskHub
 //
+//  Created by Ali Elsokary on 25/07/2026.
+//
 
 import Observation
