+++ CaskHub/Services/InstalledSoftwareScanner.swift
@@ -2,6 +2,8 @@
 //  InstalledSoftwareScanner.swift
 //  CaskHub
 //
+//  Created by Ali Elsokary on 25/07/2026.
+//
 
 import Foundation
