+++ CaskHub/Services/LocalHomebrewService+ExternalApps.swift
@@ -2,6 +2,8 @@
 //  LocalHomebrewService+ExternalApps.swift
 //  CaskHub
 //
+//  Created by Ali Elsokary on 23/07/2026.
+//
 
 import Foundation
