+++ CaskHub/Services/LocalHomebrewService+Snapshot.swift
@@ -2,6 +2,8 @@
 //  LocalHomebrewService+Snapshot.swift
 //  CaskHub
 //
+//  Created by Ali Elsokary on 25/07/2026.
+//
 
 import Foundation
