+++ CaskHub/ViewModels/CatalogProjector.swift
@@ -2,6 +2,8 @@
 //  CatalogProjector.swift
 //  CaskHub
 //
+//  Created by Ali Elsokary on 25/07/2026.
+//
 
 import Foundation
