+++ CaskHubTests/AdoptionTests.swift
@@ -337,74 +337,3 @@ final class AdoptionSurfaceTests: XCTestCase {
         XCTAssertEqual(decoded[0].keys, ["app", "zap"])
     }
 }
-
-// MARK: - View render smoke tests
-
-final class AdoptionViewRenderTests: XCTestCase {
-    @MainActor
-    private func render(_ view: some View, width: CGFloat = 420, height: CGFloat = 400) {
-        let hosting = NSHostingView(rootView: AnyView(view))
-        hosting.frame = NSRect(x: 0, y: 0, width: width, height: height)
-        hosting.layoutSubtreeIfNeeded()
-    }
-
-    @MainActor
-    func test_cask_actions_render_every_external_state() {
-        let service = LocalHomebrewService(defaults: makeScratchDefaults("render-actions"))
-        updateInstallationSnapshot(
-            of: service,
-            installedCasks: [
-                "managed": LocalCaskInstallation(
-                    token: "managed",
-                    installedVersion: "1.0",
-                    installedAt: nil,
-                    appBundleNames: ["Managed.app"]
-                )
-            ],
-            externalAppNames: ["Chrome.app"],
-            macAppStoreAppNames: ["Store.app"],
-            externalBinaryPaths: [
-                "claude": URL(fileURLWithPath: "/usr/local/bin/claude")
-            ]
-        )
-
-        let adoptable = makeCask("chrome", appNames: ["Chrome.app"])
-        render(CaskActionsView(cask: adoptable).environment(service).environment(\.isAdoptPage, true))
-        render(
-            CaskActionsView(cask: adoptable, usesIconOnlyOpenAndUpdate: true)
-                .environment(service)
-        )
-        render(CaskActionsView(cask: makeCask("store", appNames: ["Store.app"])).environment(service))
-        render(CaskActionsView(cask: makeCask("claude-code", binaryNames: ["claude"])).environment(service))
-        render(CaskActionsView(cask: makeCask("plain")).environment(service))
-        render(
-            CaskActionsView(cask: makeCask("managed", version: "2.0"), onUninstall: {})
-                .environment(service)
-        )
-    }
-
-    @MainActor
-    func test_cask_action_alerts_render_with_pending_permission_and_error() async {
-        let service = LocalHomebrewService(defaults: makeScratchDefaults("render-alerts"))
-        service.permissionProbe = { .denied }
-        try? await service.adopt(token: "chrome")
-        service.openApp(token: "chrome")
-
-        render(
-            Text("host")
-                .caskActionAlerts(for: makeCask("chrome"), showUninstallConfirmation: .constant(false))
-                .environment(service)
-        )
-    }
-
-    @MainActor
-    func test_info_popover_renders_external_and_installed_versions() {
-        let service = LocalHomebrewService(defaults: makeScratchDefaults("render-popover"))
-        render(CaskInfoPopover(cask: makeCask("mystery", appNames: ["NoSuchApp.app"])).environment(service))
-
-        updateInstalledCask(LocalCaskInstallation(
-            token: "known", installedVersion: "3.1", installedAt: .now, appBundleNames: []
-        ), in: service)
-        render(CaskInfoPopover(cask: makeCask("known")).environment(service))
-    }
-}
