+++ CaskHubTests/HomebrewCommandExecutorTests.swift
@@ -2,6 +2,8 @@
 //  HomebrewCommandExecutorTests.swift
 //  CaskHubTests
 //
+//  Created by Ali Elsokary on 25/07/2026.
+//
 
 @testable import CaskHub
 import XCTest
